+++ pkg/eib/eib.go
@@ -139,9 +139,9 @@ func buildCombustion(ctx *image.Context, rootDir string) (*combustion.Combustion
 
 		imgPath := filepath.Join(ctx.ImageConfigDir, "base-images", ctx.ImageDefinition.Image.BaseImage)
 		imgType := ctx.ImageDefinition.Image.ImageType
-		baseBuilder := resolver.NewTarballBuilder(ctx.BuildDir, imgPath, imgType, p)
+		baseBuilder := resolver.NewTarballBuilder(ctx.BuildDir, imgPath, imgType, string(ctx.ImageDefinition.Image.Arch), p)
 
-		combustionHandler.RPMResolver = resolver.New(ctx.BuildDir, p, baseBuilder, "")
+		combustionHandler.RPMResolver = resolver.New(ctx.BuildDir, p, baseBuilder, "", string(ctx.ImageDefinition.Image.Arch))
 		combustionHandler.RPMRepoCreator = rpm.NewRepoCreator(ctx.BuildDir)
 	}