+++ RELEASE_NOTES.md
@@ -14,6 +14,7 @@
 
 ## Bug Fixes
 
+* [#442](https://github.com/suse-edge/edge-image-builder/issues/442) - Only get images from specific Kubernetes objects
 * [#429](https://github.com/suse-edge/edge-image-builder/issues/429) - Automatically set execute bit on scripts
 
 ---