+++ RELEASE_NOTES.md
@@ -20,6 +20,8 @@
 
 ## Bug Fixes
 
+* [#498](https://github.com/suse-edge/edge-image-builder/issues/498) - Fix kernelArgs issue with Leap Micro 6.0
+
 ---
 
 # v1.0.2