+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@
 * Artifact sources origin and metadata are now extracted from a configuration file (`config/artifacts.yaml`)
 * Dropped `-chart` suffix from installed Helm chart names
 * Added ability to build aarch64 images on an aarch64 host machine
+* Added caching for container images
 
 ## API